+++ client/src/components/UpcomingAndHistoryInterviews/AddReviewButton.jsx
@@ -1,21 +1,15 @@
 import React, { useState } from 'react';
 import {
-  Modal, Button, Input, message, Form,
+  Modal, Button, message,
 } from 'antd';
 import './style.css';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-
-const { TextArea } = Input;
-const { Item } = Form;
+import Form from './Form';
 
 const AddReviewButton = ({ id }) => {
   const [isModalVisible, setIsModalVisible] = useState(false);
   const [review, setReview] = useState('');
-
-  const onChange = (e) => {
-    setReview(e.target.value);
-  };
 
   const showModal = () => {
     setIsModalVisible(true);
@@ -48,37 +42,7 @@ const AddReviewButton = ({ id }) => {
         onOk={handleOk}
         onCancel={handleCancel}
       >
-        <Form
-          name="basic"
-          labelCol={{
-            span: 20,
-          }}
-          wrapperCol={{
-            span: 25,
-          }}
-          initialValues={{
-            remember: true,
-          }}
-          autoComplete="off"
-        >
-          <Item
-            label="Review"
-            name="review"
-            rules={[
-              {
-                required: true,
-                message: 'Please input your review!',
-              },
-            ]}
-          >
-            <TextArea
-              className="textArea"
-              value={review}
-              placeholder="Add your review here."
-              onChange={onChange}
-            />
-          </Item>
-        </Form>
+        <Form review={review} setReview={setReview} />
       </Modal>
     </>
   );
